requirements.txt
yfinance>=0.2.40
pandas>=2.2
requests>=2.31
python-telegram-bot>=20.7
python-dotenv>=1.0
